IBX-10670: Merge IconButton and Button components
